--- ePort/src/app/account/signup/signup.component.html
@@ -1,1 +1,0 @@
-<p>signup works!</p>
