--- ePort/src/app/account/signup/signup.component.ts
@@ -1,36 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-import { Router, ActivatedRoute} from '@angular/router'
-import { FormBuilder, FormGroup, Validators } from '@angular/forms';
-
-// need to implement account and alert service
-
-@Component({
-  templateUrl: './signup.component.html',
-  styleUrls: ['./signup.component.css']
-})
-export class SignupComponent implements OnInit {
-  form!: FormGroup;
-  loading: false = false;
-  submitted: false = false; 
-  
-
-  constructor(
-    private formBuilder: FormBuilder,
-    private route: ActivatedRoute,
-    private router: Router
-  ) { }
-
-  ngOnInit(){
-    this.form = this.formBuilder.group({
-        firstName: ['', Validators.required],
-        lastName: ['', Validators.required],
-        username: ['', Validators.required],
-        password: ['', [Validators.required, Validators.minLength(6)]]
-    });
-}
-
-// convenience getter for easy access to form fields
-get f() { return this.form.controls; }
-
-
-}
