+++ ePort/src/app/app-router/app-router.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { LoginComponent } from '../login/login.component';
-import { SignupComponent } from '../account/signup/signup.component';
+import { SignupComponent } from '../signup/signup.component';
 
 
 const routes: Routes = [
@@ -11,7 +11,7 @@ const routes: Routes = [
     component: LoginComponent
   },
   {
-    path: 'account/signup',
+    path: 'signup',
     component: SignupComponent
   }
 ];
