+++ ePort/src/app/app.module.ts
@@ -5,9 +5,8 @@ import { HttpClientModule } from '@angular/common/http'
 import { AppRouterModule } from './app-router/app-router.module';
 import { AppComponent } from './app.component';
 import { LoginComponent } from './login/login.component';
-import { SignupComponent } from './account/signup/signup.component';
+import { SignupComponent } from './signup/signup.component';
 import { FormsModule } from '@angular/forms';
-import { LayoutComponent } from './account/layout/layout.component';
 import { HomeComponent } from './home/home.component';
 
 @NgModule({
@@ -15,7 +14,6 @@ import { HomeComponent } from './home/home.component';
     AppComponent,
     LoginComponent,
     SignupComponent,
-    LayoutComponent,
     HomeComponent
   ],
   imports: [
