+++ ePort/src/app/login/login.component.css
@@ -12,7 +12,7 @@ img {
 
 .login {
     text-align: center;
-}
+}     
 
 input[type=text], input[type=password] {
     width: 100%;
