+++ ePort/src/app/login/login.component.html
@@ -3,15 +3,18 @@
   <form>
     <div class="login">
       <h1>ePort</h1>
-      <img src="https://cdn.discordapp.com/attachments/975764660311711806/988891048774369410/mcd.jpeg" alt="Marshall, Carter & Dark Logo">
+      <img src="https://cdn.discordapp.com/attachments/975764660311711806/988898713575493662/mcd.png" alt="Marshall, Carter & Dark Logo">
     </div>
     
       <input type="text" name="u" placeholder="Username" required="required" />
       <input type="password" name="p" placeholder="Password" required="required" />
-      <button type="submit" class="btn btn-primary btn-block btn-large">Log-In</button>
+      <button type="submit" class="btn btn-primary btn-block btn-large">Login</button>
+      <a type="submit" routerLink="../signup" class="btn btn-link">Signup</a>
+      
       
   </form>
 </body>
 </html>
+
