+++ ePort/src/app/signup/signup.component.ts
@@ -1,0 +1,22 @@
+import { Component, OnInit } from '@angular/core';
+import { Router, ActivatedRoute} from '@angular/router'
+import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { User } from '../models/user';
+
+// need to implement account and alert service
+
+@Component({
+  selector: 'signup',
+  templateUrl: './signup.component.html',
+  styleUrls: ['./signup.component.css']
+})
+export class SignupComponent implements OnInit {
+
+  constructor(
+
+  ) { }
+
+  ngOnInit(): void {
+  }
+
+}
